feat(sub-quests): Implement basic sub-quest functionality
This commit introduces the foundational components for adding sub-quests to a quest.

- **Data & Domain Layer:**
    - Added `SubQuestEntity` and `SubQuestModel` to represent sub-quest data.
    - Created `SubQuestDao` for database operations (upsert, delete).
    - Implemented `SubQuestRepository` and its `SubQuestRepositoryImpl`.
    - Added `AddSubQuestUseCase`, `AddSubQuestsUseCase`, and `DeleteSubQuestUseCase`.
    - Integrated `SubQuestDao` into `AppDatabase`.
    - Created `SubQuestModule` for Hilt dependency injection.

- **Create Quest Screen (`CreateQuestScreen.kt`):**
    - Users can now add, view, and edit sub-tasks directly within the quest creation form.
    - A new "Unteraufgabe hinzufügen" button adds a new sub-task text field.
    - Sub-tasks are stored in the `CreateQuestUiState` and saved to the database when the quest is created.
    - Replaced an `AppTextButton` with an `AppOutlinedButton`.

- **UI & System Adjustments:**
    - Simplified the `enableEdgeToEdge` implementation in `ActivityMain.kt`.
    - Added `navigationBarsPadding` to `QuestOverviewScreen`, `AllQuestsPage`, and `QuestsForCategoryPage` to prevent UI elements from overlapping with the navigation bar.
    - Removed a commented-out PostHog feature flag check in `MainScreen.kt`.

+++ app/src/main/kotlin/de/ljz/questify/feature/quests/data/models/SubQuestEntity.kt
@@ -30,5 +30,5 @@ data class SubQuestEntity(
     val isDone: Boolean = false,
 
     @ColumnInfo(name = "quest_id")
-    val questId: Int
+    val questId: Long
 )
